backend/requirements.txt
fastapi
uvicorn
pydub
python-multipart
pillow
librosa
numpy
scipy
